#8: dedup resolveDestinationFile + extract downloads-subdir constant
