+++ composeApp/src/jvmMain/kotlin/com/tubetoast/tether/network/FileServer.kt
@@ -20,7 +20,7 @@ actual class FileServer internal constructor(
 ) {
     constructor(
         port: Int,
-        downloadsDir: File = File(System.getProperty("user.home"), "Downloads/Tether"),
+        downloadsDir: File = File(System.getProperty("user.home"), DEFAULT_DOWNLOADS_SUBDIR),
         trustedDeviceStore: TrustedDeviceStore,
         deviceKeyPair: DeviceKeyPair,
         tracker: TransferActivityTracker = DefaultTransferActivityTracker(),
@@ -85,20 +85,4 @@ private class JvmUploadStorage(
     override fun logError(message: String) {
         System.err.println("[FileServer] ERROR: $message")
     }
-}
-
-private fun resolveDestinationFile(root: File, relativePath: String): File {
-    val leafName = relativePath.substringAfterLast('/')
-    val parentPath = relativePath.substringBeforeLast('/', "")
-    val parentDir = if (parentPath.isEmpty()) root else File(root, parentPath)
-    var dest = File(parentDir, leafName)
-    if (!dest.exists()) return dest
-    val ext = leafName.substringAfterLast('.', "")
-    val base = if (ext.isEmpty()) leafName else leafName.removeSuffix(".$ext")
-    var i = 1
-    do {
-        dest = File(parentDir, if (ext.isEmpty()) "${base}_$i" else "${base}_$i.$ext")
-        i++
-    } while (dest.exists())
-    return dest
 }
